backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
razorpay==1.4.1
django-environ==0.11.2
psycopg2-binary==2.9.9
django-filter==23.3
twilio==8.10.0
gunicorn==21.2.0
whitenoise==6.6.0
